feat: implem write to version & remove manifest
