+++ public/version.json
@@ -1,0 +1,3 @@
+{
+  "head_commit": "a9a01d19b3bb141bcb253fd16565ef4e7d763bb4"
+}
